Refactor into separate component files
