+++ fastapi_server/src/rio_app/components/replay_pack_builder/name_template_component.py
@@ -1,0 +1,21 @@
+# pyright: reportUnknownMemberType=false, reportUnknownArgumentType=false, reportAttributeAccessIssue=false, reportImplicitOverride=false
+
+import rio
+
+
+class NameTemplateComponent(rio.Component):
+    replay_name_pattern: str = ""
+
+    def build(self) -> rio.Component:
+        # TODO Add tooltip
+        return rio.Column(
+            rio.Text("Name template", style="heading1"),
+            rio.Grid(
+                [rio.Text("Custom pattern"), rio.TextInput(self.bind().replay_name_pattern)],
+                # TODO Parse example replay and match for preview
+                [rio.Text("Preview"), rio.Text(self.replay_name_pattern, overflow="wrap")],
+                column_spacing=1,
+                row_spacing=1,
+                align_x=0,
+            ),
+        )
