+++ fastapi_server/src/rio_app/routes/index.py
@@ -1,10 +1,5 @@
-import time
-from collections.abc import Callable
-
-from fastapi import APIRouter, FastAPI, Request, Response
-from fastapi.routing import APIRoute
+from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
-
 
 router = APIRouter()
