+++ fastapi_server/src/routes_/audiobook/epub_upload.py
@@ -12,8 +12,6 @@ from litestar.di import Provide
 from litestar.enums import MediaType, RequestEncodingType
 from litestar.params import Body
 from litestar.response import Template
-
-from models.audiobook import AudiobookBook, AudiobookChapter
 from routes.audiobook.epub_reader import (
     EpubChapter,
     EpubMetadata,
@@ -21,6 +19,8 @@ from routes.audiobook.epub_reader import (
     extract_metadata,
 )
 from routes.cookies_and_guards import LoggedInUser, is_logged_in_guard, provide_logged_in_user
+
+from models.audiobook import AudiobookBook, AudiobookChapter
 
 load_dotenv()
