+++ fastapi_server/src/routes_/audiobook/temp_generate_tts.py
@@ -2,7 +2,6 @@ import asyncio
 import io
 
 import edge_tts
-
 from routes.caches import global_cache
