+++ fastapi_server/src/routes_/telegram_browser/cookies_and_guards.py
@@ -6,7 +6,6 @@ from dotenv import load_dotenv
 from litestar.connection import ASGIConnection
 from litestar.exceptions import NotAuthorizedException
 from litestar.handlers.base import BaseRouteHandler
-
 from routes.cookies_and_guards import COOKIES, provide_logged_in_user
 
 load_dotenv()
