+++ fastapi_server/src/routes_/telegram_browser/telegram_browser.py
@@ -23,10 +23,10 @@ from loguru import logger
 from minio import Minio
 from minio.helpers import _BUCKET_NAME_REGEX
 from pydantic import BaseModel, PositiveInt
-
-from models.telegram_browser import Status, TelegramChannel, TelegramMessage
 from routes.caches import cache_coroutine_result
 from routes.telegram_browser.cookies_and_guards import is_logged_in_allowed_accounts_guard
+
+from models.telegram_browser import Status, TelegramChannel, TelegramMessage
 
 BUCKET_NAME = os.getenv("MINIO_TELEGRAM_FILES_BUCKET")
 assert BUCKET_NAME is not None
