+++ fastapi_server/src/routes_/text_to_speech.py
@@ -19,7 +19,6 @@ from litestar.enums import RequestEncodingType
 from litestar.params import Body
 from litestar.response import Template
 from pydantic import BaseModel
-
 from routes.tts.generate_tts import Voices, generate_tts
 
 WS_BACKEND_SERVER_URL = os.getenv("BACKEND_WS_SERVER_URL", "ws:0.0.0.0:8000")
