+++ fastapi_server/src/routes_/tts/irc_bot_async.py
@@ -7,7 +7,6 @@ from typing import TYPE_CHECKING
 
 from httpx_ws import AsyncWebSocketSession, aconnect_ws
 from loguru import logger
-
 from routes.tts.generate_tts import Voices
 
 if TYPE_CHECKING:
