+++ fastapi_server/src/routes_/tts/irc_bot_async2.py
@@ -18,7 +18,6 @@ from collections.abc import Callable
 from typing import Literal
 
 from loguru import logger
-
 from routes.tts.generate_tts import Voices
 
 # pyrefly: ignore
